Extract profile helper to shared module
