+++ integration-tests/api-tests/spec/ingredient.spec.js
@@ -1,5 +1,5 @@
 const executeRequest = require('./helpers/request-helper');
-const ProfileHelper = require('./helpers/profile-helper');
+const ProfileHelper = require('../../shared/profile-helper');
 
 describe('Ingredient', () => {
     beforeAll(() => {
