+++ integration-tests/shared/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "integration-tests-shared",
+  "version": "1.0.0",
+  "license": "ISC",
+  "devDependencies": {
+    "sqlite3": "^5.1.4"
+  }
+}
