+++ integration-tests/shared/profile-helper/profiles/simple-recipe.json
