+++ integration-tests/shared/profile-helper/profiles/test-recipe.json
